Separate XML escaping from using entities for non-encodable chars
The '&', '<', '>' and sometimes quotes always have to be escaped
because they have have a special meaning in XML. Whether other
characters have to be escaped depends on what encoding is used
when converting the XML to bytes, so it should be done at that
time, not when generating the XML.

Note that this also removes the mapping of non-printable ASCII
characters to spaces. I couldn't find any reason for this conversion.