Determine whether a document is XML or HTML using content type header
This might be too simplistic, but it's better than checking for the
string "xmlns=", which was the previous approach.

Perhaps in the future we could also look at doctype declarations.

In any case, determining the type outside of parse_document() is an
improvement, since the code that detects the encoding also needs to
know whether the document is XML or not.