Perform validation after parsing
The spider needs a parsed document whether it is valid or not.
Instead of parsing with validation first and falling back on more
lenient parsers if necessary, we now parse with recovery enabled
from the start. lxml will report errors that are recovered from,
so we will not lose any strictness there.

This also enables validation for documents that needed recovery.
I think this is an advantage, since we can report more problems
in a single run, but if it turns out to produce too many false
positives we might have to disable it.

The code is much more modular now, since parsing and validation
was split over two functions.

Another advantage is that we can cache the DTD instead of loading
it again for every document we parse.

It is also easier to do more sophisticated validation in the future,
such as using other validation methods such as XMLSchema or RelaxNG,
or doing validation of inline SVG content.