Change report interface to be more like a logger
Keep the messages in document order instead of grouping them by
severity. This will be useful when switching to the validator.nu
checker, which can issue an 'info' message as a follow-up to an
error or warning, for example showing the first definition location
when a duplicate definition is found.

Perhaps it makes sense to use the actual logging.Logger interface
for reports, but I have no immediate plans for that.