Tweak command line syntax
Use all-caps for non-literal values in positional arguments as well.

Make properties file argument mandatory, to avoid the first positional
argument to be mistaken for the optional plugin argument. Even though
argparse understands the '--' separator to disambiguate situations
like that, that feature is not documented. Also it's a complication
that I don't think is worth the slightly shorter syntax.

Remove spaces around '|'.