Check all documents for an XML declaration
Previously, we would only check for an XML declaration if the document
was served with an XML content type. However, it is possible the web
server or the local file type autodetection gets the content type
wrong.

ElementTree will refuse to parse documents in string form that have
an encoding specified in the XML declaration. To avoid that problem,
we remove the XML declaration if present, but this safeguard was
skipped if the content type indicated the document was not XML.