Look at encoding in XML declaration before deciding encoding
Previously, we only looked at the XML declaration if none of the
other encoding hints worked. This would lead to false positives
when a document contains only ASCII characters, since it would
then successfully decode as UTF-8, but might declare a different
encoding in the XML declaration, which would then be reported
as a mismatch.