Send local files to plugins with improved metadata
The improvised HTTP header we get for local files are less accurate
than the metadata we gather by examining the content. So instead of
sending the improvised header, for local files we construct a new
header before we pass the content to plugins.

This avoids a lot of false positives when using the v.Nu plugin
on local files.