Look for URLs in multiple XML namespaces
Previously we only searched for URLs in XHTML attributes. But it is
possible to embed SVG in XHTML, XHTML in Atom etc., so it makes more
sense to check every node for every namespace, regardless of the
root namespace.

With this new approach, we now find URLs in SVG, Atom and XLink.