Fix local file handling
This got broken when the ability to check the content of redirects
was added, because the fake response object we used for local
redirects was not complete enough. Instead of improving the fake
response, I opted for acting more like urllib works internally,
since that should be more future proof.