Do not check content for non-200 responses
This seemed like an interesting feature, but after testing it in
practice I found that both Nginx and Twisted use HTML in their
redirect bodies that doesn't pass the HTML5 checker. So having
this enabled by default will cause a lot of false positives.

It would be interesting to offer this as an option that the user
can turn on, but we don't have an easy framework yet for enabling
or disabling checks.