Integrate IncrementalReport functionality into Report
The 'url' variable exists on the base class, so there is no reason to
not always insert it into the log record.

The formatting of lxml error objects is harmless to have in the base
class, since it is not used when a 'str' is passed. Besides, this code
should be moved out, so it's not a reason to keep IncrementalReport
around in the long run.