Require Python 3.6
I want to add type annotations for new code and in the long term also
annotate the existing code. The Python 3.5 syntax for annotating
variables and fields is not very convenient, so I'd rather use the
3.6 syntax from the start.

Additionally, this makes it easier to share Invoke tasks between APE
and SoftFab.

Also bump the version to reflect this incompatible change.