Raise ValueError instead of UnicodeDecodeError if no codec matches
UnicodeDecodeError is only suitable for reporting that a particular
decode attempt failed. The way our code tried to call it won't work
because it's missing a lot of constructor arguments.

This issue was found using mypy. The type annotations have been
stashed to allow some refactoring, but will appear soon.