Undo incorrect change made when adding type annotations
In an attempt to reduce the amount of code, I accidentally also reduced
the scope of the 'try' statement in a way that breaks the code.

While I did test APE after the change, for some reason I haven't
figured out yet, an older version of APE was run instead of the
updated version until I exited the Poetry shell and re-entered it.