Decode namespace prefix if it is stored as bytes
Flagged by mypy.

In practice the namespace seems to be stored as a string when APE
runs, but it is not clear to me whether this is always the case
under Python 3 or whether it depends on what data we pass to lxml.
To be on the safe side, we check the type and decode bytes if
necessary.