Install lxml-stubs package
This is almost a PEP 561 compliant stubs package, so we can install
it as a package instead of using a git submodule.

For now, I'm using our fork, since some fixes are required to make it
work. But in the future I hope we'll be able to switch to upstream or
to PyPI.