Check whether we are allowed to spider the initial URL
It is possible for the initial URL to be disallowed by `robots.txt`,
in which case we should not request it. The previous implementation
only checked whether we are allowed to spider discovered URLs, not
the initial URL. This commit changes that by using `add_referrer()`
for the initial URL as well.

Closes #26