Update mypy to 0.920
Since `disallow_any_generics` is enabled in `mypy.ini`, the
`LoggerAdapter` type must be specialized: an implicit `Any`
is not allowed. However, at runtime `LoggerAdapter` is not
subscriptable, so we can't specialize it in code that runs.
The solution is to define some helper types differently
depending on whether we're type checking or running.